Fix requerimiento: Para escanear QR hay que apretar 'escanear' en la pantalla principal y luego volver a apretar otro boton que dice 'escanear'.
